\name{ADPopSelApp}
\alias{ADPopSelApp}
\title{
Graphical user interface to design an adaptive trial with data-driven population selection
}
\description{
This function creates a web application with a Shiny-based graphical user interface to facilitate the process of designing Phase III adaptive trials with two interim analysis to enable a futility assessment and data-driven selection of the most promising patient population or populations. The application supports the same functionality as the \code{\link{ADPopSel}} function. For an example of the function call, see \link{ADPopSelExample1}.
}
\usage{
ADPopSelApp()
}
\seealso{
  \code{\link{ADPopSel}}
}
