\name{ADRandApp}
\alias{ADRandApp}
\title{
Graphical user interface to design an adaptive trial with data-driven population selection
}
\description{
This function creates a web application with a Shiny-based graphical user interface to facilitate the process of designing dose-finding Phase II trials with response-adaptive randomization. The application supports the same functionality as the \code{\link{ADRand}} function. For an example of the function call, see \link{ADRandExample}.
}
\arguments{No arguments}
\value{No return value, called for side effects}
\usage{
ADRandApp()
}
\seealso{
  \code{\link{ADRand}}
}
