\name{ADSSModExample2}
\alias{ADSSModExample2}
\title{
Simulation-based design of an adaptive trial with sample size re-estimation (binary endpoint)
}
\description{
Consider a seamless Phase II/Phase III or Phase III trial where the primary efficacy endpoint is binary. A single dose or regimen of an experimental treatment will be compared to a control. An adaptive design with two interim analyses will be employed in the trial (the first interim analysis may correspond to the end of the Phase II portion of the trial). The following decision rules will be applied at the interim looks:
\itemize{
	\item Futility stopping rule will be applied at the first interim analysis (IA1). The trial may be stopped due to futility if a significant treatment effect is unlikely to be established at the final analysis (FA).
    \item Sample size re-estimation rule will be applied at the second interim analysis (IA2). The total number of patients in the trial will be increased to boost the probability of success if conditional power is lower than expected at this interim analysis.
} 
The following design parameters will be assumed:
\itemize{
    \item A trial design with a 1:2 randomization ratio will be used. A hundred patients will be enrolled in the control arm and twice as many patients will be enrolled in the experimental treatment arm.
    \item The patient dropout rate at the end of the treatment period is equal to 10\%.
    \item The information fractions at IA1 and IA2 are set to 40\% and 60\%. 
}
In addition, the following decision rules will be considered at the two interim analyses:
\itemize{
    \item The futility threshold at IA1 is set to 20\%.
    \item The promising interval at IA2 ranges from 50\% to 90\%.
    \item The target conditional power at IA2 is set to 90\% and the total sample size in the trial could be increased up to 40\%.
}
The response rates for the primary efficacy endpoint are assumed to be 10\% (control arm)and 25\% (experimental treatment arm).

Key operating characteristics of this adaptive design will be evaluated using the \code{\link{ADSSMod}} function with 10,000 simulation runs. To invoke this function, a list of all design and decision rule parameters (\code{parameters}) needs to be created as shown below. A comprehensive simulation report can be generated by calling the \code{\link{GenerateReport}} function and a graphical user interface can be launched using the \code{\link{ADSSModApp}} function.
}
\seealso{
  \code{\link{ADSSMod}}, \code{\link{ADSSModExample1}}, \code{\link{ADSSModExample3}}
}
\examples{
# List of all parameters
parameters = list()

# Endpoint type
parameters$endpoint_type = "Binary"

# Number of enrolled patients (control, treatment) 
parameters$sample_size = c(100, 200)

# Patient dropout rate
parameters$dropout_rate = 0.1

# Response rate in the control arm 
parameters$control_rate = 0.1

# Response rate in the treatment arm 
parameters$treatment_rate = 0.25

# Information fractions at IA1, IA2, FA (before sample size adjustment) 
# and FA (after sample size adjustment)
parameters$info_frac = c(0.4, 0.6, 1, 1.4)

# Futility threshold for conditional power at IA1
parameters$futility_threshold = 0.2

# Promising interval for conditional power at IA2
parameters$promising_interval = c(0.5, 0.9)

# Target conditional power for increasing the sample size at IA2
parameters$target_power = 0.9

# One-sided alpha level
parameters$alpha = 0.025

# Number of simulations
parameters$nsims = 10000

# Run simulations to compute operating characteristics
results = ADSSMod(parameters)

# Generate a simulation report (remove tempfile)
GenerateReport(results, 
              tempfile("ADSSMod Binary endpoint.docx", fileext=".docx"))
}