\name{ADTreatSelApp}
\alias{ADTreatSelApp}
\title{
Graphical user interface to design an adaptive trial with data-driven treatment selection
}
\description{
This function creates a web application with a Shiny-based graphical user interface to facilitate the process of designing Phase III adaptive trials with two interim analysis to enable a futility assessment and data-driven selection of the best treatment arm. The application supports the same functionality as the \code{\link{ADTreatSel}} function. For an example of the function call, see \link{ADTreatSelExample1}.
}
\usage{
ADTreatSelApp()
}
\seealso{
  \code{\link{ADTreatSel}}
}
