\name{ClustRandApp}
\alias{ClustRandApp}
\title{
Graphical user interface to design a cluster-randomized trial
}
\description{
This function creates a web application with a Shiny-based graphical user interface to facilitate the process of designing a two-arm cluster-randomized trial (treatment and control arms). The application supports the same functionality as the \code{\link{ClustRand}} function. For an example of the function call, see \link{ClustRandExample1}.
}
\arguments{No arguments}
\value{No return value, called for side effects}
\usage{
ClustRandApp()
}
\seealso{
  \code{\link{ClustRand}}
}
