\name{MultAdjApp}
\alias{MultAdjApp}
\title{
Graphical user interface for power calculations in traditional trials with multiple objectives
}
\description{
This function creates a web application with a Shiny-based graphical user interface to perform power calculations in traditional trials with multiple objectives. The application supports the same functionality as the \code{\link{MultAdj}} function. For an example of the function call, see \link{MultAdjExample1}.
}
\arguments{No arguments}
\value{No return value, called for side effects}
\usage{
MultAdjApp()
}
\seealso{
  \code{\link{MultAdj}}, \code{\link{MultAdj}}, \code{\link{MultAdjExample2}}, \code{\link{MultAdjExample3}}
}
